product 내 관리자 파트 분할, 이동
