+++ backend/src/test/java/com/back/domain/admin/controller/AdminProductControllerTest.java
@@ -1,7 +1,6 @@
-package com.back.domain.product.controller;
+package com.back.domain.admin.controller;
 
 import com.back.domain.product.dto.product.ProductRequestDto;
-import com.back.domain.product.dto.product.ProductSearchDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -24,7 +23,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @AutoConfigureMockMvc
 @ActiveProfiles("test")
 @Transactional
-public class ProductControllerTest {
+public class AdminProductControllerTest {
     
     @Autowired
     private MockMvc mockMvc;
@@ -56,110 +55,6 @@ public class ProductControllerTest {
                 "유효하지 않은 상품 설명",
                 1
         );
-    }
-
-    // ========== 공개 API 테스트 (인증 불필요) ==========
-
-    @Test
-    @DisplayName("GET /api/v1/products - 상품 목록 조회 (인증 불필요)")
-    @WithMockUser
-    void getAllProducts_Success() throws Exception {
-        mockMvc.perform(get("/api/v1/products"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
-    }
-
-    @Test
-    @DisplayName("GET /api/v1/products/{productId} - 상품 상세 조회 성공 (인증 불필요)")
-    @WithMockUser
-    void getProductById_Success() throws Exception {
-        mockMvc.perform(get("/api/v1/products/{productId}", 1))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id").value(1))
-                .andExpect(jsonPath("$.name").exists());
-    }
-
-    @Test
-    @DisplayName("GET /api/v1/products/{productId} - 상품 상세 조회 실패 (상품 없음)")
-    @WithMockUser
-    void getProductById_NotFound() throws Exception {
-        mockMvc.perform(get("/api/v1/products/{productId}", 9999))
-                .andDo(print())
-                .andExpect(status().isNotFound());
-    }
-
-    // ========== 검색 API 테스트 (인증 불필요) ==========
-
-    @Test
-    @DisplayName("POST /api/v1/products/search - 통합 검색 (인증 불필요)")
-    @WithMockUser
-    void searchProducts_Success() throws Exception {
-        ProductSearchDto searchDto = new ProductSearchDto(
-                "에티오피아", null, null, null, null, true, false
-        );
-
-        mockMvc.perform(post("/api/v1/products/search")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(searchDto)))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
-    }
-
-    @Test
-    @DisplayName("GET /api/v1/products/search - 쿼리 파라미터 검색 (인증 불필요)")
-    @WithMockUser
-    void searchProductsWithParams_Success() throws Exception {
-        mockMvc.perform(get("/api/v1/products/search")
-                        .param("name", "콜롬비아"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
-    }
-
-    @Test
-    @DisplayName("GET /api/v1/products/search/name - 상품명 검색 (인증 불필요)")
-    @WithMockUser
-    void searchProductsByName_Success() throws Exception {
-        mockMvc.perform(get("/api/v1/products/search/name")
-                        .param("name", "브라질"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
-    }
-
-    @Test
-    @DisplayName("GET /api/v1/products/category/{categoryId} - 카테고리별 상품 조회 (인증 불필요)")
-    @WithMockUser
-    void getProductsByCategory_Success() throws Exception {
-        mockMvc.perform(get("/api/v1/products/category/{categoryId}", 2))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
-    }
-
-    @Test
-    @DisplayName("GET /api/v1/products/price-range - 가격 범위 검색 (인증 불필요)")
-    @WithMockUser
-    void getProductsByPriceRange_Success() throws Exception {
-        mockMvc.perform(get("/api/v1/products/price-range")
-                        .param("minPrice", "16000")
-                        .param("maxPrice", "25000"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
-    }
-
-    @Test
-    @DisplayName("GET /api/v1/products/out-of-stock - 품절 상품 조회 (인증 불필요)")
-    @WithMockUser
-    void getOutOfStockProducts_Success() throws Exception {
-        mockMvc.perform(get("/api/v1/products/out-of-stock"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
     }
 
     // ========== 관리자 전용 API 테스트 ==========
@@ -252,5 +147,112 @@ public class ProductControllerTest {
                 .andExpect(status().isNotFound());
     }
 
+    @Test
+    @DisplayName("PUT /api/v1/admin/products/{productId} - 상품 수정 성공 (관리자 권한)")
+    @WithMockUser(roles = "ADMIN")
+    void updateProduct_Success_WithAdmin() throws Exception {
+        ProductRequestDto updateDto = new ProductRequestDto(
+                "수정된 상품",
+                "http://updated.com/image.jpg",
+                20000,
+                45,
+                "수정된 상품 설명입니다.",
+                2
+        );
 
+        mockMvc.perform(put("/api/v1/admin/products/{productId}", 1)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(updateDto)))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.id").value(1))
+                .andExpect(jsonPath("$.name").value("수정된 상품"))
+                .andExpect(jsonPath("$.price").value(20000));
+    }
+
+    @Test
+    @DisplayName("PUT /api/v1/admin/products/{productId} - 상품 수정 실패 (일반 사용자)")
+    @WithMockUser(roles = "USER")
+    void updateProduct_Forbidden_WithUser() throws Exception {
+        mockMvc.perform(put("/api/v1/admin/products/{productId}", 1)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(testProductRequestDto)))
+                .andDo(print())
+                .andExpect(status().isForbidden());
+    }
+
+    @Test
+    @DisplayName("PUT /api/v1/admin/products/{productId} - 상품 수정 실패 (상품 없음)")
+    @WithMockUser(roles = "ADMIN")
+    void updateProduct_NotFound() throws Exception {
+        mockMvc.perform(put("/api/v1/admin/products/{productId}", 9999)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(testProductRequestDto)))
+                .andDo(print())
+                .andExpect(status().isNotFound());
+    }
+
+    @Test
+    @DisplayName("PUT /api/v1/admin/products/{productId} - 상품 수정 실패 (유효성 검증 실패)")
+    @WithMockUser(roles = "ADMIN")
+    void updateProduct_BadRequest_Validation() throws Exception {
+        mockMvc.perform(put("/api/v1/admin/products/{productId}", 1)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(invalidProductRequestDto)))
+                .andDo(print())
+                .andExpect(status().isBadRequest());
+    }
+
+    @Test
+    @DisplayName("DELETE /api/v1/admin/products/{productId} - 상품 삭제 성공 (관리자 권한)")
+    @WithMockUser(roles = "ADMIN")
+    void deleteProduct_Success_WithAdmin() throws Exception {
+        // 새로운 상품을 먼저 생성하고 삭제하는 방식으로 테스트
+        String response = mockMvc.perform(post("/api/v1/admin/products")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(testProductRequestDto)))
+                .andExpect(status().isCreated())
+                .andReturn()
+                .getResponse()
+                .getContentAsString();
+        
+        // ObjectMapper를 사용하여 안전하게 ID 추출
+        var responseNode = objectMapper.readTree(response);
+        int createdId = responseNode.get("id").asInt();
+        
+        // 생성된 상품 삭제 - HTTP Status 204 No Content 응답 확인
+        mockMvc.perform(delete("/api/v1/admin/products/{productId}", createdId))
+                .andDo(print())
+                .andExpect(status().isNoContent());
+
+        // 삭제 확인 - 공개 API로 확인
+        mockMvc.perform(get("/api/v1/products/{productId}", createdId))
+                .andExpect(status().isNotFound());
+    }
+
+    @Test
+    @DisplayName("DELETE /api/v1/admin/products/{productId} - 상품 삭제 실패 (일반 사용자)")
+    @WithMockUser(roles = "USER")
+    void deleteProduct_Forbidden_WithUser() throws Exception {
+        mockMvc.perform(delete("/api/v1/admin/products/{productId}", 1))
+                .andDo(print())
+                .andExpect(status().isForbidden());
+    }
+
+    @Test
+    @DisplayName("DELETE /api/v1/admin/products/{productId} - 상품 삭제 실패 (인증 없음)")
+    void deleteProduct_Unauthorized_WithoutAuth() throws Exception {
+        mockMvc.perform(delete("/api/v1/admin/products/{productId}", 1))
+                .andDo(print())
+                .andExpect(status().isUnauthorized());
+    }
+
+    @Test
+    @DisplayName("DELETE /api/v1/admin/products/{productId} - 상품 삭제 실패 (상품 없음)")
+    @WithMockUser(roles = "ADMIN")
+    void deleteProduct_NotFound() throws Exception {
+        mockMvc.perform(delete("/api/v1/admin/products/{productId}", 9999))
+                .andDo(print())
+                .andExpect(status().isNotFound());
+    }
 }
